Adding the last functions part 1 of x
New in this version:
* CSV file shows all requested columns in header now
* Only Global class is static now
** Each class that needs only a single instantiation gets instantiated
by Globals
** Globals passes a reference to itself to the constructor of that
class, to prevent never ending recursion
** When a project is re-opened, the Global class is created, the
experiment class is read and references to the global class are then
updated in that class
* Navigating to trials and looks within those trials is now possible
through the menu
* Saving as is now possible through the menu
* When opening a class, a check is executed to see if the file still
exists. If not, the user is asked to re-select it
* When saving, a check is now executed to see if a file with the same
name already exists. If so, the user is asked if they want to overwrite
that file or cancel

What still needs to be done:
* More general cleaning up
* The settings view within the program switches experiment name and
experiment ID, fix this
* The PlayerControlsPanel now has a static function to format
milliseconds to a nice string. Implement this throughout the code
* Implement all Todo's
* Handle exceptions and stop printing to the terminal
* Fix when trials and looks can be started and ended, and make sure the
buttons only allow to do that when the software allows it as well (Maybe
use a function that checks if something is valid, that is both used to
enable or disable the buttons, as by the experiment model before it does
the check, because now, those are two separate functions)