Button and if adding succeeds now depend on same function
The model is now converted to two abstract classes.
Experiment, Trial and Look implement those classes.

This way, each of these classes work the same, with a lot of basic
functions being used by each of the non-abstract classes and working
exactly the same at that.

The buttons are now being disabled and enabled using the same test
function to check if creating or ending a trial or look is allowed,
making sure the buttons appear if and only if the action that button
executes is allowed.

Bugs:
- Once a trial or look is created, the endtime is set to -1 to show
something needs to be done with that. At this point, after the first
trial is created, all buttons get grayed out, presumably because this
end-time is never updated. This is something that needs to be fixed