Previous frame works for all files except mpg
Previous frame function now works for all files except mpg. This file
extension is removed from files that can be opened.

The trouble is that VLCJ Library jumps to a certain time stamp and then
jumps a few seconds ahead still after a few seconds. This 'offset' can
be recorded and adjusted for automatically, as now is happening.