Read on to see why this user downgraded to an older version
Basically, just because 3.1.0 requires VLC 2.1.x at the least, but
Ubuntu 12.04 LTS only goes up to VLC 2.0.0 (which can be upgraded, it
just is a lot of trouble, which is probably not such a good idea).

Seeing as all the PC's in the lab run some version of Ubuntu 12, trying
to get it to work for VLC 2.0.0 seems to be the best option.

Will keep you posted!