Added functionality and corrected adding of items bug
Through the menu trials and looks can now be deleted. Menu items can
only be clicked when the functions can actually be executed.

Furthermore, a bug still rested in the adding of new items. When adding
a a new item before another item, making it the new *first* item, the
program would not allow it. This has now been solved.
The begin- and end time of items is now corrected for the items
surrounding it, so no items on the same level overlap in time anymore