Stage 1 of allowing for alternative VLC installation paths
This application requires VLC. To run, it must locate the install
directory of VLC and load a few of its libraries. In some cases, VLC may
not be found in the locations where one would expect to find it. In
these cases, it is of the utmost vital importance that the user can
select the custom installation directory. This, however, does not work
because exceptions in Java seem to do something with the JNI or JNA
library, whatever it's called.

So this is pretty cool.
#not