CSV Exporter and keyDispatch
CSV Exporter now uses quotes around column names to indicate them being
just a single column.

CSV Exporter now shows the total look time in a trial, instead of the
total duration of the trial itself (this was a bug).

KeyDispatch has been changed to KeyListener, so the video doesn't try to
do all kinds of things when a user is entering a name for a save
location for example