Tune in this commit to learn if anything new ever really happens
The application now works both under 64bit Java & VLC and 32bit Java &
VLC on Windows and it also works on Java (either version) on Linux. This
was achieved by reverting to a previous JNA version (from 4.1.0 to
3.5.2).
In the original problem, JNA couldn't find the plugins of VLC which were
located in exactly the same directory as the DLL's, but appearantly that
has something to do with JNA.

When a file is saved, the name and location of the file are now saved
separately, because what with the file seperators on different systems
stuff didn't really work. Unfortunately, this breaks every file created
before this commit (solution: try saving it with a different name and
location).