VLC build in bug now differently worked around
A bug in VLC jumps the time to a different time than the time that it is
told to jump to. This 'error' is constant, so I tried to work around
this problem by adding an offset equal to the error.
This was done by jumping to 1000ms and then checking the actual time.
The difference would be the offset or error.

But now I got the problem that the offset seemed to be off as well. This
turned out to be because the error was about 2500ms, meaning it can
never be guessed correct.
To solve this problem, the offset is calculated by jumping to the middle
of the video and checking the actual time against the expected time. The
difference is now the offset.

This should work for any offset, unless the offset is larger than half
the media length, in which case you definitely shouldn't use that video
anyway.